requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiohttp_pydantic==2.4.0
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.3.0
colorama==0.4.6
exceptiongroup==1.3.0
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.6.4
openapi-generator-cli==7.16.0
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
pytest==8.2.1
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
swagger_ui_bundle==1.1.0
tomli==2.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
